DATAMONGO-1968 - Add configuration support for com.mongodb.client.MongoClient
