+++ spring-data-mongodb/src/main/java/org/springframework/data/mongodb/MongoDbFactory.java
@@ -58,6 +58,11 @@ public interface MongoDbFactory extends CodecRegistryProvider {
 	 */
 	PersistenceExceptionTranslator getExceptionTranslator();
 
+	/**
+	 * Get the legacy database entry point. Please consider {@link #getDb()} instead.
+	 *
+	 * @return
+	 */
 	DB getLegacyDb();
 
 	/**
